requirements.txt
Tortoise-ORM==0.25.1
fastapi==0.116.1
uvicorn[standard]==0.27.0
mysqlclient==2.2.7
celery[redis]==5.5.0
redis==5.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
pydantic==2.4.0
requests==2.31.0
black==24.1.0
flake8==6.1.0
passlib[bcrypt]==1.7.4
# Fixed bcrypt version to prevent password length errors in hashing (72 byte limit)
bcrypt==4.0.1
python-jose==3.5.0
authlib==1.5.1
aiomysql==0.2.0
tensorflow==2.20.0
scikit-learn==1.7.2
numpy==2.3.4